requirements.txt

langchain
langchain-google-genai
google-generativeai
gradio
